+++ src-docs/src/views/code/code_example.js
@@ -62,6 +62,8 @@ export default props => (
             <li><EuiCode>paddingSize</EuiCode> accepts "s" / "m" / "l" (default).</li>
             <li><EuiCode>fontSize</EuiCode> accepts "s" (default) / "m" / "l".</li>
             <li><EuiCode>overflowHeight</EuiCode> accepts a number. By default it is not set.</li>
+            <li><EuiCode>transparentBackground</EuiCode> set to <EuiCode>false</EuiCode> will remove the background.</li>
+            <li><EuiCode>inline</EuiCode>Will display the passed code in an inline format.</li>
           </ul>
         </div>
       }