+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [`master`](https://github.com/elastic/eui/tree/master)
 
 - Changed the hover states of `EuiButtonEmpty` to look more like links.
+- Added `inline` and `transparentBackground` props to `EuiCodeBlock`. Made light theme the default.
 
 # [`0.0.1`](https://github.com/elastic/eui/tree/v0.0.1) Initial Release